Split type-testing and assignment snippets
