+++ csharp/language-reference/operators/TypeTestingAndConversionOperators.cs
@@ -3,63 +3,15 @@ using System.Collections.Generic;
 
 namespace operators
 {
-    public static class TypeTestingAndAssignmentOperators
+    public static class TypeTestingAndConversionOperators
     {
         public static void Examples()
         {
-            Assignment();
-            RefAssignment();
             Cast();
             IsOperatorExample.Main();
             IsOperatorWithInt();
             IsOperatorTypePattern();
             AsOperator();
-        }
-
-        private static void Assignment()
-        {
-            // <SnippetAssignment>
-            var numbers = new List<double>() { 1.0, 2.0, 3.0 };
-
-            Console.WriteLine(numbers.Capacity);
-            numbers.Capacity = 100;
-            Console.WriteLine(numbers.Capacity);
-            // Output:
-            // 4
-            // 100
-
-            int newFirstElement;
-            double originalFirstElement = numbers[0];
-            newFirstElement = 5;
-            numbers[0] = newFirstElement;
-            Console.WriteLine(originalFirstElement);
-            Console.WriteLine(numbers[0]);
-            // Output:
-            // 1
-            // 5
-            // </SnippetAssignment>
-        }
-
-        private static void RefAssignment()
-        {
-            // <SnippetRefAssignment>
-            void Display(double[] s) => Console.WriteLine(string.Join(" ", s));
-            
-            double[] arr = { 0.0, 0.0, 0.0 };
-            Display(arr);
-
-            ref double arrayElement = ref arr[0];
-            arrayElement = 3.0;
-            Display(arr);
-
-            arrayElement = ref arr[arr.Length - 1];
-            arrayElement = 5.0;
-            Display(arr);
-            // Output:
-            // 0 0 0
-            // 3 0 0
-            // 3 0 5
-            // </SnippetRefAssignment>
         }
 
         private static void Cast()
